Updated index.html, style.css and README
Changes made to index.html were classses added and divs reorganised to use grid instead of float for positioning.
Changes made to style.css were grid and flex layout desiogn and start of work on checkboxes to look like switches.
Readme updates with the steps takes so far on project